delete topic on hard delete

+++ dataproduct-registry/src/main/java/io/confluent/dpregistry/web/rest/resources/DataProductRegistryResource.java
@@ -398,7 +398,9 @@ public class DataProductRegistryResource extends SchemaRegistryResource {
       dataProductRegistry.deleteDataProductOrForward(
           cluster, env, name, permanentDelete, headerProperties);
 
-      deleteTopic(env, cluster, product.value);
+      if (permanentDelete) {
+        deleteTopic(env, cluster, product.value);
+      }
 
       asyncResponse.resume(Response.status(204).build());
     } catch (SchemaRegistryException e) {
